# back_end/school-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API SCHOOL",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://school-app.test/api"
        }
    ],
    "paths": {
        "/admin/course": {
            "get": {
                "tags": [
                    "course"
                ],
                "summary": "listar",
                "description": "Listar cursos",
                "operationId": "getcourse",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Internal Server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/course/register": {
            "post": {
                "tags": [
                    "course"
                ],
                "summary": "Register course",
                "description": "Registrar al Cursos",
                "operationId": "getCourseRegister",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nombre",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_start",
                        "in": "query",
                        "description": "fecha de inicio",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_end",
                        "in": "query",
                        "description": "fecha de fin",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "schedule",
                        "in": "query",
                        "description": "horario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "tipo de modalidad",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/admin/course/{id}": {
            "get": {
                "tags": [
                    "course"
                ],
                "summary": "obtener",
                "description": "Obtener cursos",
                "operationId": "getCourseId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "el id del curso",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "delete": {
                "tags": [
                    "course"
                ],
                "summary": "Eliminar",
                "description": "Eliminar cursos",
                "operationId": "getCourseDeleteId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "el id del curso",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/admin/course/updated": {
            "post": {
                "tags": [
                    "course"
                ],
                "summary": "Actualizar course",
                "description": "Actualizar al Cursos",
                "operationId": "getCourseUpdated",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nombre",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_start",
                        "in": "query",
                        "description": "fecha de inicio",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_end",
                        "in": "query",
                        "description": "fecha de fin",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "schedule",
                        "in": "query",
                        "description": "horario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "tipo de modalidad",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Iinternal Serve"
                    }
                }
            }
        },
        "/admin/dashboard": {
            "get": {
                "tags": [
                    "dashboard"
                ],
                "summary": "listar",
                "description": "Listar Tablero",
                "operationId": "getDashboard",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Internal Server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    }
}